♻️  Refactor games into a custom hook
